requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==1.10.13
python-dotenv==1.0.1
requests==2.31.0
langchain>=0.0.354
langchain-groq>=0.0.1
langchain-community>=0.0.38
langchain-core>=0.1.5,<0.2
langchain-text-splitters>=0.0.2
langsmith>=0.1.0,<0.2.0
typing-extensions==4.9.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pymongo==4.6.1
PyJWT==2.8.0 